code refactor in icons folder
